+++ pom.xml
@@ -23,7 +23,7 @@
         <kotlin.version>1.6.10</kotlin.version>
         <mockk.version>1.12.3</mockk.version>
         <confluent.version>6.2.1</confluent.version>
-        <prosessering.version>1.20211125150647_a7c6f64</prosessering.version>
+        <prosessering.version>1.20220308102416_2479ea0</prosessering.version>
         <felles.version>1.20220221162703_957e034</felles.version>
         <kontrakter.version>2.0_20220222213922_64795e8</kontrakter.version>
         <main-class>no.nav.familie.ba.mottak.LauncherKt</main-class>
