+++ src/main/kotlin/no/nav/familie/ba/mottak/domene/Hendelseslogg.kt
@@ -1,14 +1,10 @@
 package no.nav.familie.ba.mottak.domene
 
-import java.io.IOException
-import java.io.StringReader
-import java.io.StringWriter
+import no.nav.familie.ba.mottak.util.PropertiesToStringConverter
 import java.time.LocalDateTime
 import java.util.Properties
-import javax.persistence.AttributeConverter
 import javax.persistence.Column
 import javax.persistence.Convert
-import javax.persistence.Converter
 import javax.persistence.Entity
 import javax.persistence.EnumType
 import javax.persistence.Enumerated
@@ -53,34 +49,4 @@ enum class HendelseConsumer {
         JOURNAL_AIVEN,
         EF_VEDTAK_V1,
         EF_VEDTAK_INFOTRYGD_V1
-}
-
-/**
- * JPA konverterer for å skrive ned en key=value text til et databasefelt (output tilsvarer java.util.Properties
- * format).
- */
-@Converter
-class PropertiesToStringConverter : AttributeConverter<Properties, String> {
-
-        override fun convertToDatabaseColumn(props: Properties?): String? {
-                if (props == null || props.isEmpty) {
-                        return null
-                }
-                val stringWriter = StringWriter(512)
-                // custom i stedet for Properties.store slik at vi ikke får med default timestamp
-                props.forEach { key, value -> stringWriter.append(key as String).append('=').append(value as String).append('\n') }
-                return stringWriter.toString()
-        }
-
-        override fun convertToEntityAttribute(dbData: String?): Properties {
-                val props = Properties()
-                if (dbData != null) {
-                        try {
-                                props.load(StringReader(dbData))
-                        } catch (e: IOException) {
-                                throw IllegalArgumentException("Kan ikke lese properties til string:$props", e) //$NON-NLS-1$
-                        }
-                }
-                return props
-        }
 }
